requirements.txt
flet==0.21.2
psycopg[binary]
python-dotenv==1.0.1
